Move search specific strings to seatch module
